# feat: Vercel Blob storage for gallery image optimization